Rename sp_response_ext -> sp_response_expect
